refactor: rename AllServicesDefinition to Implementation and streamline configuration handling
